Renamed all models to use Extended suffix, put suitable classes as inner
